requirements.txt
functions-framework==3.10.1
google-cloud-bigquery==3.18.0
google-cloud-storage==1.32.0
pyarrow==21.0.0
requests==2.31.0
